Moved the ffi part into a separate library
