--- pysrc/otto/__init__.py
@@ -1,5 +1,0 @@
-"""Library for GPGPU
-
-*otto* is a library for GPGPU built in C using OpenCL. This is a wrapper for this
-library, built using cffi.
-"""
