--- pysrc/otto/defs.py
@@ -1,10 +1,0 @@
-"""Basic definitions for the library"""
-
-from pathlib import Path
-
-
-BASE_PATH = Path(__file__).parent
-ROOT_PATH = BASE_PATH.parent.parent
-
-PYBUILD_DIRNAME = "pysrc"
-PYBUILD_PATH = Path.joinpath(ROOT_PATH, PYBUILD_DIRNAME)
