+++ pysrc/otto_ffi/config/__init__.py
@@ -1,7 +1,12 @@
 from pathlib import Path
 
-from otto.defs import ROOT_PATH
 
+BASE_PATH = Path(__file__).parent.parent
+ROOT_PATH = BASE_PATH.parent.parent
+SHLIB_PATH = BASE_PATH
+
+PYBUILD_DIRNAME = "pysrc"
+PYBUILD_PATH = Path.joinpath(ROOT_PATH, PYBUILD_DIRNAME)
 
 DEBUG_DIRNAME = "debug"
 
@@ -12,6 +17,15 @@ CBIN_DEBUG_PATH = Path.joinpath(ROOT_PATH, "bin", DEBUG_DIRNAME)
 CLIB_RELEASE_PATH = Path.joinpath(ROOT_PATH, "lib")
 CLIB_DEBUG_PATH = Path.joinpath(ROOT_PATH, "lib", DEBUG_DIRNAME)
 CSOURCES_PATH = Path.joinpath(ROOT_PATH, "src")
+
+CLIB_SO_NAMES = [
+    "libotto_core.so",
+    "libotto_utils.so",
+]
+CLIB_DLL_NAMES = [
+    "otto_core.dll",
+    "otto_utils.dll",
+]
 
 INCLUDE_DIRS = [
     str(Path.joinpath(CSOURCES_PATH)),
