+++ setup.cfg
@@ -20,10 +20,9 @@ classifiers =
 # These options can be changed according to your project
 [options]
 packages =
-    otto
-    otto.ffi
-    otto.ffi.cl
-    otto.ffi.config
+    otto_ffi
+    otto_ffi.cl
+    otto_ffi.config
     otto_utils
 python_requires = >=3.7
 package_dir =
@@ -31,5 +30,8 @@ package_dir =
 zip_safe = no
 
 [options.package_data]
-otto = py.typed
+otto_ffi =
+    py.typed
+    *.dll
+    *.so
 otto_utils = py.typed
